python-scripts/requirements.txt
supabase==2.4.0
pydantic==2.10.2
typer==0.12.5
redis==5.2.0
requests==2.32.3
python-dotenv==1.0.1
pytest==8.3.3
